requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
python-dotenv==1.0.1